requirements.txt
# Core Dependencies
torch>=2.0.0
numpy>=1.24.0
gymnasium[atari]>=0.29.0
ale-py>=0.8.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
opencv-python>=4.8.0

# Logging and Monitoring
tensorboard>=2.13.0
tqdm>=4.65.0

# Configuration
pyyaml>=6.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0

# Utilities
pillow>=10.0.0
scipy>=1.11.0
pandas>=2.0.0
